requirements.txt
pyTelegramBotAPI==4.12.0
pillow==10.4.0
